Reorganized the project files
